+++ dockerfiles/worker
@@ -30,7 +30,8 @@ RUN rm -rf /var/lib/apt/lists/*
 COPY --from=builder /go/src/app/bin/vc_${SERVICE_NAME} /vc_service
 COPY --from=builder /go/src/app/docs /docs
 COPY --from=builder /go/src/app/standards /standards
-COPY --from=builder /go/src/app/users_paris.xlsx /users_paris.xlsx
+#COPY --from=builder /go/src/app/users_paris.xlsx /users_paris.xlsx
+COPY --from=builder /go/src/app/bootstrapping /bootstrapping
 
 EXPOSE 8080
