+++ internal/issuer/apiv1/handlers.go
@@ -7,7 +7,7 @@ import (
 	"vc/internal/gen/registry/apiv1_registry"
 	"vc/pkg/education"
 	"vc/pkg/helpers"
-	"vc/pkg/model"
+	"vc/pkg/identity"
 	"vc/pkg/socialsecurity"
 
 	"google.golang.org/grpc"
@@ -119,8 +119,8 @@ func (c *Client) MakeSDJWT(ctx context.Context, req *CreateCredentialRequest) (*
 			return nil, err
 		}
 
-	case "pid":
-		doc := &model.PIDDocument{}
+	case "PID":
+		doc := &identity.PIDDocument{}
 		if err := json.Unmarshal(req.DocumentData, &doc); err != nil {
 			return nil, err
 		}
