+++ internal/mockas/bootstrapper/method_pid.go
@@ -1,411 +1,91 @@
 package bootstrapper
 
-type pidUsers struct {
-	firstName  string
-	familyName string
-	BirthDate  string
+import (
+	"context"
+	"encoding/json"
+	"fmt"
+	"os"
+	"path/filepath"
+	"vc/pkg/identity"
+	"vc/pkg/model"
+)
+
+type pidClient struct {
+	client         *Client
+	documents      map[string]*model.CompleteDocument
+	credentialType string
 }
 
-func (c *Client) MakePID() {
-	c.pid["70"] = pidUsers{
-		firstName:  "Robert",
-		familyName: "De Niro",
-		BirthDate:  "1982/01/15",
-	}
-	c.pid["71"] = pidUsers{
-		firstName:  "Judi",
-		familyName: "Dench",
-		BirthDate:  "1995/09/04",
-	}
-	c.pid["72"] = pidUsers{
-		firstName:  "Johnny",
-		familyName: "Depp",
-		BirthDate:  "1985/12/03",
-	}
-	c.pid["73"] = pidUsers{
-		firstName:  "Leonardo",
-		familyName: "DiCaprio",
-		BirthDate:  "1972/08/28",
-	}
-	c.pid["74"] = pidUsers{
-		firstName:  "Robert",
-		familyName: "Downey Jr.",
-		BirthDate:  "1983/05/12",
-	}
-	c.pid["75"] = pidUsers{
-		firstName:  "Adam",
-		familyName: "Driver",
-		BirthDate:  "1988/09/27",
-	}
-	c.pid["76"] = pidUsers{
-		firstName:  "Idris",
-		familyName: "Elba",
-		BirthDate:  "1980/09/02",
-	}
-	c.pid["77"] = pidUsers{
-		firstName:  "Colin",
-		familyName: "Farrell",
-		BirthDate:  "1977/04/25",
-	}
-	c.pid["78"] = pidUsers{
-		firstName:  "Michael",
-		familyName: "Fassbender",
-		BirthDate:  "1992/04/24",
-	}
-	c.pid["79"] = pidUsers{
-		firstName:  "Jodie",
-		familyName: "Foster",
-		BirthDate:  "1991/10/30",
-	}
-	c.pid["80"] = pidUsers{
-		firstName:  "Jamie",
-		familyName: "Foxx",
-		BirthDate:  "1977/07/03",
-	}
-	c.pid["81"] = pidUsers{
-		firstName:  "Morgan",
-		familyName: "Freeman",
-		BirthDate:  "1990/08/24",
-	}
-	c.pid["82"] = pidUsers{
-		firstName:  "Donald",
-		familyName: "Glover",
-		BirthDate:  "2007/03/17",
-	}
-	c.pid["83"] = pidUsers{
-		firstName:  "Tom",
-		familyName: "Hanks",
-		BirthDate:  "1989/12/04",
-	}
-	c.pid["84"] = pidUsers{
-		firstName:  "Zoe",
-		familyName: "Saldana",
-		BirthDate:  "1981/11/23",
-	}
-	c.pid["85"] = pidUsers{
-		firstName:  "Anne",
-		familyName: "Hathaway",
-		BirthDate:  "1987/05/06",
-	}
-	c.pid["86"] = pidUsers{
-		firstName:  "Salma",
-		familyName: "Hayek",
-		BirthDate:  "1983/08/27",
-	}
-	c.pid["87"] = pidUsers{
-		firstName:  "Chris",
-		familyName: "Hemsworth",
-		BirthDate:  "1992/01/19",
-	}
-	c.pid["88"] = pidUsers{
-		firstName:  "Anthony",
-		familyName: "Hopkins",
-		BirthDate:  "2006/08/16",
-	}
-	c.pid["89"] = pidUsers{
-		firstName:  "Hugh",
-		familyName: "Jackman",
-		BirthDate:  "1972/06/28",
-	}
-	c.pid["90"] = pidUsers{
-		firstName:  "Samuel L.",
-		familyName: "Jackson",
-		BirthDate:  "2007/01/27",
-	}
-	c.pid["91"] = pidUsers{
-		firstName:  "Scarlett",
-		familyName: "Johansson",
-		BirthDate:  "1980/05/12",
-	}
-	c.pid["92"] = pidUsers{
-		firstName:  "Dakota",
-		familyName: "Johnson",
-		BirthDate:  "1999/04/21",
-	}
-	c.pid["93"] = pidUsers{
-		firstName:  "Nicole",
-		familyName: "Kidman",
-		BirthDate:  "1983/03/15",
-	}
-	c.pid["94"] = pidUsers{
-		firstName:  "Vanessa",
-		familyName: "Kirby",
-		BirthDate:  "1993/06/28",
-	}
-	c.pid["95"] = pidUsers{
-		firstName:  "Mila",
-		familyName: "Kunis",
-		BirthDate:  "1989/01/01",
-	}
-	c.pid["96"] = pidUsers{
-		firstName:  "Jennifer",
-		familyName: "Lawrence",
-		BirthDate:  "2006/02/14",
-	}
-	c.pid["97"] = pidUsers{
-		firstName:  "Rachel",
-		familyName: "McAdams",
-		BirthDate:  "1998/06/09",
-	}
-	c.pid["98"] = pidUsers{
-		firstName:  "Matthew",
-		familyName: "McConaughey",
-		BirthDate:  "1977/09/09",
-	}
-	c.pid["99"] = pidUsers{
-		firstName:  "Ewan",
-		familyName: "McGregor",
-		BirthDate:  "1991/08/29",
-	}
-	c.pid["100"] = pidUsers{
-		firstName:  "Helen",
-		familyName: "Mirren",
-		BirthDate:  "1996/01/30",
-	}
-	c.pid["101"] = pidUsers{
-		firstName:  "Jason",
-		familyName: "Momoa",
-		BirthDate:  "1983/08/20",
-	}
-	c.pid["102"] = pidUsers{
-		firstName:  "Gary",
-		familyName: "Oldman",
-		BirthDate:  "1988/09/09",
-	}
-	c.pid["103"] = pidUsers{
-		firstName:  "Al",
-		familyName: "Pacino",
-		BirthDate:  "1984/06/11",
-	}
-	c.pid["104"] = pidUsers{
-		firstName:  "Michelle",
-		familyName: "Pfeiffer",
-		BirthDate:  "1992/12/15",
-	}
-	c.pid["105"] = pidUsers{
-		firstName:  "Joaquin",
-		familyName: "Phoenix",
-		BirthDate:  "1979/10/16",
-	}
-	c.pid["106"] = pidUsers{
-		firstName:  "Brad",
-		familyName: "Pitt",
-		BirthDate:  "1974/12/07",
-	}
-	c.pid["107"] = pidUsers{
-		firstName:  "Natalie",
-		familyName: "Portman",
-		BirthDate:  "2005/09/01",
-	}
-	c.pid["108"] = pidUsers{
-		firstName:  "Chris",
-		familyName: "Pratt",
-		BirthDate:  "1974/01/07",
-	}
-	c.pid["109"] = pidUsers{
-		firstName:  "Keanu",
-		familyName: "Reeves",
-		BirthDate:  "1985/11/26",
-	}
-	c.pid["110"] = pidUsers{
-		firstName:  "John C.",
-		familyName: "Reilly",
-		BirthDate:  "2005/11/19",
-	}
-	c.pid["111"] = pidUsers{
-		firstName:  "Ryan",
-		familyName: "Reynolds",
-		BirthDate:  "1988/12/05",
-	}
-	c.pid["112"] = pidUsers{
-		firstName:  "Margot",
-		familyName: "Robbie",
-		BirthDate:  "1991/02/14",
-	}
-	c.pid["113"] = pidUsers{
-		firstName:  "Julia",
-		familyName: "Roberts",
-		BirthDate:  "1982/07/30",
-	}
-	c.pid["114"] = pidUsers{
-		firstName:  "Paul",
-		familyName: "Rudd",
-		BirthDate:  "2006/03/28",
-	}
-	c.pid["115"] = pidUsers{
-		firstName:  "Will",
-		familyName: "Smith",
-		BirthDate:  "1992/03/13",
-	}
-	c.pid["116"] = pidUsers{
-		firstName:  "Octavia",
-		familyName: "Spencer",
-		BirthDate:  "1997/11/18",
-	}
-	c.pid["117"] = pidUsers{
-		firstName:  "Kristen",
-		familyName: "Stewart",
-		BirthDate:  "1993/07/17",
-	}
-	c.pid["118"] = pidUsers{
-		firstName:  "Emma",
-		familyName: "Stone",
-		BirthDate:  "1985/01/27",
-	}
-	c.pid["119"] = pidUsers{
-		firstName:  "Meryl",
-		familyName: "Streep",
-		BirthDate:  "2005/07/29",
-	}
-	c.pid["120"] = pidUsers{
-		firstName:  "Tilda",
-		familyName: "Swinton",
-		BirthDate:  "1984/04/14",
-	}
-	c.pid["121"] = pidUsers{
-		firstName:  "Channing",
-		familyName: "Tatum",
-		BirthDate:  "2006/10/22",
-	}
-	c.pid["122"] = pidUsers{
-		firstName:  "Charlize",
-		familyName: "Theron",
-		BirthDate:  "1988/12/11",
-	}
-	c.pid["123"] = pidUsers{
-		firstName:  "Vince",
-		familyName: "Vaughn",
-		BirthDate:  "1988/02/19",
-	}
-	c.pid["124"] = pidUsers{
-		firstName:  "Alicia",
-		familyName: "Vikander",
-		BirthDate:  "1985/11/09",
-	}
-	c.pid["125"] = pidUsers{
-		firstName:  "Christoph",
-		familyName: "Waltz",
-		BirthDate:  "1986/03/23",
-	}
-	c.pid["126"] = pidUsers{
-		firstName:  "Denzel",
-		familyName: "Washington",
-		BirthDate:  "1981/10/28",
-	}
-	c.pid["127"] = pidUsers{
-		firstName:  "Naomi",
-		familyName: "Watts",
-		BirthDate:  "1990/07/24",
-	}
-	c.pid["128"] = pidUsers{
-		firstName:  "Michelle",
-		familyName: "Williams",
-		BirthDate:  "1992/03/29",
-	}
-	c.pid["129"] = pidUsers{
-		firstName:  "Owen",
-		familyName: "Wilson",
-		BirthDate:  "2007/09/11",
-	}
-	c.pid["130"] = pidUsers{
-		firstName:  "Kate",
-		familyName: "Winslet",
-		BirthDate:  "1986/06/19",
-	}
-	c.pid["131"] = pidUsers{
-		firstName:  "Reese",
-		familyName: "Witherspoon",
-		BirthDate:  "1987/04/02",
-	}
-	c.pid["132"] = pidUsers{
-		firstName:  "Amy",
-		familyName: "Adams",
-		BirthDate:  "1990/10/05",
-	}
-	c.pid["133"] = pidUsers{
-		firstName:  "Ben",
-		familyName: "Affleck",
-		BirthDate:  "1973/07/05",
-	}
-	c.pid["134"] = pidUsers{
-		firstName:  "Anne",
-		familyName: "Bancroft",
-		BirthDate:  "1980/12/30",
-	}
-	c.pid["135"] = pidUsers{
-		firstName:  "Halle",
-		familyName: "Berry",
-		BirthDate:  "1989/08/25",
-	}
-	c.pid["136"] = pidUsers{
-		firstName:  "Jessica",
-		familyName: "Biel",
-		BirthDate:  "1988/09/12",
-	}
-	c.pid["137"] = pidUsers{
-		firstName:  "Cate",
-		familyName: "Blanchett",
-		BirthDate:  "2006/11/08",
-	}
-	c.pid["138"] = pidUsers{
-		firstName:  "Emily",
-		familyName: "Blunt",
-		BirthDate:  "2000/08/15",
-	}
-	c.pid["139"] = pidUsers{
-		firstName:  "John",
-		familyName: "Boyega",
-		BirthDate:  "2005/12/05",
-	}
-	c.pid["140"] = pidUsers{
-		firstName:  "Sandra",
-		familyName: "Bullock",
-		BirthDate:  "1984/09/08",
-	}
-	c.pid["141"] = pidUsers{
-		firstName:  "Nicolas",
-		familyName: "Cage",
-		BirthDate:  "2005/05/13",
-	}
-	c.pid["142"] = pidUsers{
-		firstName:  "Steve",
-		familyName: "Carell",
-		BirthDate:  "2005/06/04",
-	}
-	c.pid["143"] = pidUsers{
-		firstName:  "Jessica",
-		familyName: "Chastain",
-		BirthDate:  "1982/10/19",
-	}
-	c.pid["144"] = pidUsers{
-		firstName:  "Glenn",
-		familyName: "Close",
-		BirthDate:  "1981/02/05",
-	}
-	c.pid["145"] = pidUsers{
-		firstName:  "Daniel",
-		familyName: "Craig",
-		BirthDate:  "1988/10/29",
-	}
-	c.pid["146"] = pidUsers{
-		firstName:  "Penélope",
-		familyName: "Cruz",
-		BirthDate:  "1994/05/22",
-	}
-	c.pid["147"] = pidUsers{
-		firstName:  "Benedict",
-		familyName: "Cumberbatch",
-		BirthDate:  "1987/10/24",
-	}
-	c.pid["148"] = pidUsers{
-		firstName:  "Matt",
-		familyName: "Damon",
-		BirthDate:  "1971/07/09",
-	}
-	c.pid["149"] = pidUsers{
-		firstName:  "Viola",
-		familyName: "Davis",
-		BirthDate:  "1985/08/11",
+func NewPIDClient(ctx context.Context, client *Client) (*pidClient, error) {
+	pda1Client := &pidClient{
+		client:         client,
+		documents:      map[string]*model.CompleteDocument{},
+		credentialType: "pid",
 	}
 
+	return pda1Client, nil
+}
+
+func (c *pidClient) makeSourceData(sourceFilePath string) error {
+	for pidNumber, id := range c.client.identities {
+		c.documents[pidNumber] = &model.CompleteDocument{}
+
+		documentData := identity.PIDDocument{
+			GivenName:  id.Identities[0].GivenName,
+			FamilyName: id.Identities[0].FamilyName,
+			BirthDate:  id.Identities[0].BirthDate,
+		}
+
+		var err error
+		c.documents[pidNumber].DocumentData, err = documentData.Marshal()
+		if err != nil {
+			return err
+		}
+
+		c.documents[pidNumber].Meta = &model.MetaData{
+			AuthenticSource: "",
+			DocumentVersion: "1.0.0",
+			DocumentType:    "PID",
+			DocumentID:      fmt.Sprintf("document_id_pda1_%s", pidNumber),
+			RealData:        false,
+			Collect: &model.Collect{
+				ID:         fmt.Sprintf("collect_id_pda1_%s", pidNumber),
+				ValidUntil: 0,
+			},
+			Revocation:                &model.Revocation{},
+			CredentialValidFrom:       0,
+			CredentialValidTo:         0,
+			DocumentDataValidationRef: "",
+		}
+
+		c.documents[pidNumber].DocumentDisplay = &model.DocumentDisplay{
+			Version: "1.0.0",
+			Type:    "secure",
+			DescriptionStructured: map[string]any{
+				"en": map[string]any{
+					"documentType": "PID",
+				},
+				"sv": map[string]any{
+					"documentType": "PID",
+				},
+			},
+		}
+	}
+
+	return nil
+}
+
+func (c *pidClient) save2Disk() error {
+	b, err := json.MarshalIndent(c.documents, "", "  ")
+	if err != nil {
+		return err
+	}
+
+	filePath := filepath.Join("../../../bootstrapping", fmt.Sprintf("%s.json", c.credentialType))
+
+	if err := os.WriteFile(filePath, b, 0644); err != nil {
+		return err
+	}
+
+	return nil
 }
